rp/gpio: implement Input
